Procfile/requirements.txt
altair==5.5.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comtypes==1.4.8
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.0
fonttools==4.59.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.9
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
mysql-connector-python==9.4.0
narwhals==1.15.2
networkx==3.4.2
numpy==2.1.3
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==5.29.1
pyarrow==18.1.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.3
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.98
pytz==2024.2
pywin32==308
pyzbar==0.1.9
qrcode==8.0
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
selenium==4.27.1
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
streamlit==1.40.2
tenacity==9.0.0
textfsm==2.1.0
toml==0.10.2
tornado==6.4.2
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
